+++ tests/Tests/Unit/GapicClientTraitTest.php
@@ -41,6 +41,7 @@ use Google\ApiCore\CredentialsWrapper;
 use Google\ApiCore\GapicClientTrait;
 use Google\ApiCore\LongRunning\OperationsClient;
 use Google\ApiCore\OperationResponse;
+use Google\ApiCore\Options\TransportOptions;
 use Google\ApiCore\RequestParamsHeaderDescriptor;
 use Google\ApiCore\RetrySettings;
 use Google\ApiCore\ServerStream;
@@ -71,7 +72,7 @@ class GapicClientTraitTest extends TestCase
     public function tearDown(): void
     {
         // Reset the static gapicVersion field between tests
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('gapicVersionFromFile', null, true);
     }
 
@@ -119,7 +120,7 @@ class GapicClientTraitTest extends TestCase
                     'credentialsWrapper' => $credentialsWrapper,
                 ])
             );
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('agentHeader', $header);
         $client->set('retrySettings', [
             'method' => $this->getMockBuilder(RetrySettings::class)
@@ -130,39 +131,33 @@ class GapicClientTraitTest extends TestCase
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('descriptors', ['method' => $unaryDescriptors]);
-        $client->call('startApiCall', [
+        $client->startApiCall(
             'method',
             $request,
             ['headers' => $headers]
-        ]);
+        );
     }
 
     public function testConfigureCallConstructionOptionsAcceptsRetryObjectOrArray()
     {
         $defaultRetrySettings = RetrySettings::constructDefault();
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('retrySettings', ['method' => $defaultRetrySettings]);
         $expectedOptions = [
             'retrySettings' => $defaultRetrySettings
                 ->with(['rpcTimeoutMultiplier' => 5])
         ];
-        $actualOptionsWithObject = $client->call(
-            'configureCallConstructionOptions',
+        $actualOptionsWithObject = $client->configureCallConstructionOptions(
+            'method',
             [
-                'method',
-                [
-                    'retrySettings' => $defaultRetrySettings
-                        ->with(['rpcTimeoutMultiplier' => 5])
-                ]
+                'retrySettings' => $defaultRetrySettings
+                    ->with(['rpcTimeoutMultiplier' => 5])
             ]
         );
-        $actualOptionsWithArray = $client->call(
-            'configureCallConstructionOptions',
+        $actualOptionsWithArray = $client->configureCallConstructionOptions(
+            'method',
             [
-                'method',
-                [
-                    'retrySettings' => ['rpcTimeoutMultiplier' => 5]
-                ]
+                'retrySettings' => ['rpcTimeoutMultiplier' => 5]
             ]
         );
 
@@ -193,7 +188,7 @@ class GapicClientTraitTest extends TestCase
              ->method('startUnaryCall')
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -205,12 +200,12 @@ class GapicClientTraitTest extends TestCase
             ->setMethodsExcept(['validate'])
             ->getMock();
 
-        $response = $client->call('startOperationsCall', [
+        $response = $client->startOperationsCall(
             'method',
             [],
             $message,
             $operationsClient
-        ])->wait();
+        )->wait();
 
         $expectedResponse = new OperationResponse(
             '',
@@ -245,7 +240,7 @@ class GapicClientTraitTest extends TestCase
              ->method('startUnaryCall')
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitOperationsStub();
+        $client = new OperationsGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -258,10 +253,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('operationsClient', $operationsClient);
 
         $request = new MockRequest();
-        $response = $client->call('startApiCall', [
+        $response = $client->startApiCall(
             'method',
             $request
-        ])->wait();
+        )->wait();
 
         $expectedResponse = new OperationResponse(
             '',
@@ -297,7 +292,7 @@ class GapicClientTraitTest extends TestCase
              ->method('startUnaryCall')
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitOperationsStub();
+        $client = new OperationsGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -310,10 +305,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('operationsClient', $operationsClient);
 
         $request = new MockRequest();
-        $response = $client->call('startApiCall', [
+        $response = $client->startApiCall(
             'method',
             $request,
-        ])->wait();
+        )->wait();
 
         $expectedResponse = new OperationResponse(
             '',
@@ -329,7 +324,7 @@ class GapicClientTraitTest extends TestCase
      */
     public function testStartApiCallException($descriptor, $expected)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('descriptors', $descriptor);
 
         // All descriptor config checks throw Validation exceptions
@@ -337,10 +332,10 @@ class GapicClientTraitTest extends TestCase
         // Check that the proper exception is being thrown for the given descriptor.
         $this->expectExceptionMessage($expected);
 
-        $client->call('startApiCall', [
+        $client->startApiCall(
             'method',
             new MockRequest()
-        ])->wait();
+        )->wait();
     }
 
     public function startApiCallExceptions()
@@ -406,7 +401,7 @@ class GapicClientTraitTest extends TestCase
             )
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -414,10 +409,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('descriptors', ['method' => $unaryDescriptors]);
 
         $request = new MockRequest();
-        $client->call('startApiCall', [
+        $client->startApiCall(
             'method',
             $request
-        ])->wait();
+        )->wait();
     }
 
     public function testStartApiCallPaged()
@@ -444,7 +439,7 @@ class GapicClientTraitTest extends TestCase
              ->method('startUnaryCall')
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -452,10 +447,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('descriptors', ['method' => $pagedDescriptors]);
 
         $request = new MockRequest();
-        $client->call('startApiCall', [
+        $client->startApiCall(
             'method',
             $request
-        ]);
+        );
     }
 
     public function testStartAsyncCall()
@@ -474,7 +469,7 @@ class GapicClientTraitTest extends TestCase
              ->method('startUnaryCall')
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -482,10 +477,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('descriptors', ['Method' => $unaryDescriptors]);
 
         $request = new MockRequest();
-        $client->call('startAsyncCall', [
+        $client->startAsyncCall(
             'method',
             $request
-        ])->wait();
+        )->wait();
     }
 
     public function testStartAsyncCallPaged()
@@ -519,7 +514,7 @@ class GapicClientTraitTest extends TestCase
             )
              ->will($this->returnValue($expectedPromise));
         $credentialsWrapper = CredentialsWrapper::build([]);
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -527,10 +522,10 @@ class GapicClientTraitTest extends TestCase
         $client->set('descriptors', ['Method' => $pagedDescriptors]);
 
         $request = new MockRequest();
-        $client->call('startAsyncCall', [
+        $client->startAsyncCall(
             'method',
             $request
-        ])->wait();
+        )->wait();
     }
 
     /**
@@ -538,7 +533,7 @@ class GapicClientTraitTest extends TestCase
      */
     public function testStartAsyncCallException($descriptor, $expected)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('descriptors', $descriptor);
 
         // All descriptor config checks throw Validation exceptions
@@ -546,10 +541,10 @@ class GapicClientTraitTest extends TestCase
         // Check that the proper exception is being thrown for the given descriptor.
         $this->expectExceptionMessage($expected);
 
-        $client->call('startAsyncCall', [
+        $client->startAsyncCall(
             'method',
             new MockRequest()
-        ])->wait();
+        )->wait();
     }
 
     public function startAsyncCallExceptions()
@@ -600,24 +595,24 @@ class GapicClientTraitTest extends TestCase
         require_once __DIR__ . '/testdata/src/GapicClientStub.php';
         $version = '1.2.3-dev';
         $client = new \GapicClientStub();
-        $this->assertEquals($version, $client->call('getGapicVersion', [[]]));
+        $this->assertEquals($version, $client->getGapicVersion([]));
     }
 
     public function testGetGapicVersionWithNoAvailableVersion()
     {
-        $client = new GapicClientTraitStub();
-        $this->assertSame('', $client->call('getGapicVersion', [[]]));
+        $client = new StubGapicClient();
+        $this->assertSame('', $client->getGapicVersion([]));
     }
 
     public function testGetGapicVersionWithLibVersion()
     {
         $version = '1.2.3-dev';
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('gapicVersionFromFile', $version, true);
         $options = ['libVersion' => $version];
-        $this->assertEquals($version, $client->call('getGapicVersion', [
+        $this->assertEquals($version, $client->getGapicVersion(
             $options
-        ]));
+        ));
     }
 
     /**
@@ -625,11 +620,11 @@ class GapicClientTraitTest extends TestCase
      */
     public function testCreateCredentialsWrapper($auth, $authConfig, $expectedCredentialsWrapper)
     {
-        $client = new GapicClientTraitStub();
-        $actualCredentialsWrapper = $client->call('createCredentialsWrapper', [
+        $client = new StubGapicClient();
+        $actualCredentialsWrapper = $client->createCredentialsWrapper(
             $auth,
             $authConfig,
-        ]);
+        );
 
         $this->assertEquals($expectedCredentialsWrapper, $actualCredentialsWrapper);
     }
@@ -654,14 +649,14 @@ class GapicClientTraitTest extends TestCase
      */
     public function testCreateCredentialsWrapperValidationException($auth, $authConfig)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         $this->expectException(ValidationException::class);
 
-        $client->call('createCredentialsWrapper', [
+        $client->createCredentialsWrapper(
             $auth,
             $authConfig,
-        ]);
+        );
     }
 
     public function createCredentialsWrapperValidationExceptionData()
@@ -677,14 +672,14 @@ class GapicClientTraitTest extends TestCase
      */
     public function testCreateCredentialsWrapperInvalidArgumentException($auth, $authConfig)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         $this->expectException(InvalidArgumentException::class);
 
-        $client->call('createCredentialsWrapper', [
+        $client->createCredentialsWrapper(
             $auth,
             $authConfig,
-        ]);
+        );
     }
 
     public function createCredentialsWrapperInvalidArgumentExceptionData()
@@ -702,12 +697,12 @@ class GapicClientTraitTest extends TestCase
         if ($expectedTransportClass == GrpcTransport::class) {
             $this->requiresGrpcExtension();
         }
-        $client = new GapicClientTraitStub();
-        $transport = $client->call('createTransport', [
+        $client = new StubGapicClient();
+        $transport = $client->createTransport(
             $apiEndpoint,
             $transport,
             $transportConfig
-        ]);
+        );
 
         $this->assertEquals($expectedTransportClass, get_class($transport));
     }
@@ -739,15 +734,15 @@ class GapicClientTraitTest extends TestCase
      */
     public function testCreateTransportInvalid($apiEndpoint, $transport, $transportConfig)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         $this->expectException(ValidationException::class);
 
-        $client->call('createTransport', [
+        $client->createTransport(
             $apiEndpoint,
             $transport,
             $transportConfig
-        ]);
+        );
     }
 
     public function createTransportDataInvalid()
@@ -768,12 +763,12 @@ class GapicClientTraitTest extends TestCase
 
     public function testServiceAddressAlias()
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $apiEndpoint = 'test.address.com:443';
-        $updatedOptions = $client->call('buildClientOptions', [
+        $updatedOptions = $client->buildClientOptions(
             ['serviceAddress' => $apiEndpoint]
-        ]);
-        $client->call('setClientOptions', [$updatedOptions]);
+        );
+        $client->setClientOptions($updatedOptions);
 
         $this->assertEquals($apiEndpoint, $updatedOptions['apiEndpoint']);
         $this->assertArrayNotHasKey('serviceAddress', $updatedOptions);
@@ -782,14 +777,14 @@ class GapicClientTraitTest extends TestCase
     public function testOperationClientClassOption()
     {
         $options = ['operationsClientClass' => CustomOperationsClient::class];
-        $client = new GapicClientTraitStub();
-        $operationsClient = $client->call('createOperationsClient', [$options]);
+        $client = new StubGapicClient();
+        $operationsClient = $client->createOperationsClient($options);
         $this->assertInstanceOf(CustomOperationsClient::class, $operationsClient);
     }
 
     public function testAdditionalArgumentMethods()
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         // Set the LRO descriptors we are testing.
         $longRunningDescriptors = [
@@ -831,12 +826,12 @@ class GapicClientTraitTest extends TestCase
         $operationsClient->getOperation('test-123', 'abc', 100)
             ->shouldBeCalledOnce();
 
-        $operationResponse = $client->call('startOperationsCall', [
+        $operationResponse = $client->startOperationsCall(
             'method.name',
             [],
             $request,
             $operationsClient->reveal()
-        ])->wait();
+        )->wait();
 
         // This will invoke $operationsClient->getOperation with values from
         // the additional argument methods.
@@ -848,9 +843,9 @@ class GapicClientTraitTest extends TestCase
      */
     public function testSetClientOptions($options, $expectedProperties)
     {
-        $client = new GapicClientTraitStub();
-        $updatedOptions = $client->call('buildClientOptions', [$options]);
-        $client->call('setClientOptions', [$updatedOptions]);
+        $client = new StubGapicClient();
+        $updatedOptions = $client->buildClientOptions($options);
+        $client->setClientOptions($updatedOptions);
         foreach ($expectedProperties as $propertyName => $expectedValue) {
             $actualValue = $client->get($propertyName);
             $this->assertEquals($expectedValue, $actualValue);
@@ -859,7 +854,7 @@ class GapicClientTraitTest extends TestCase
 
     public function setClientOptionsData()
     {
-        $clientDefaults = GapicClientTraitStub::getClientDefaults();
+        $clientDefaults = StubGapicClient::getClientDefaults();
         $expectedRetrySettings = RetrySettings::load(
             $clientDefaults['serviceName'],
             json_decode(file_get_contents($clientDefaults['clientConfig']), true)
@@ -872,7 +867,7 @@ class GapicClientTraitTest extends TestCase
         }
         $expectedProperties = [
             'serviceName' => 'test.interface.v1.api',
-            'agentHeader' => AgentHeader::buildAgentHeader([]),
+            'agentHeader' => AgentHeader::buildAgentHeader([]) + ['User-Agent' => ['gcloud-php-legacy/']],
             'retrySettings' => $expectedRetrySettings,
         ];
         return [
@@ -889,8 +884,8 @@ class GapicClientTraitTest extends TestCase
         if (!extension_loaded('sysvshm')) {
             $this->markTestSkipped('The sysvshm extension must be installed to execute this test.');
         }
-        $client = new GapicClientTraitStub();
-        $updatedOptions = $client->call('buildClientOptions', [$options]);
+        $client = new StubGapicClient();
+        $updatedOptions = $client->buildClientOptions($options);
         $this->assertEquals($expectedUpdatedOptions, $updatedOptions);
     }
 
@@ -971,8 +966,8 @@ class GapicClientTraitTest extends TestCase
         if (!extension_loaded('sysvshm')) {
             $this->markTestSkipped('The sysvshm extension must be installed to execute this test.');
         }
-        $client = new GapicClientTraitRestOnly();
-        $updatedOptions = $client->call('buildClientOptions', [$options]);
+        $client = new RestOnlyGapicClient();
+        $updatedOptions = $client->buildClientOptions($options);
         $this->assertEquals($expectedUpdatedOptions, $updatedOptions);
     }
 
@@ -1036,8 +1031,8 @@ class GapicClientTraitTest extends TestCase
      */
     public function testBuildRequestHeaders($headerParams, $request, $expected)
     {
-        $client = new GapicClientTraitStub();
-        $actual = $client->call('buildRequestParamsHeader', [$headerParams, $request]);
+        $client = new StubGapicClient();
+        $actual = $client->buildRequestParamsHeader($headerParams, $request);
         $this->assertEquals($actual[RequestParamsHeaderDescriptor::HEADER_KEY], $expected);
     }
 
@@ -1117,14 +1112,14 @@ class GapicClientTraitTest extends TestCase
     public function testModifyClientOptions()
     {
         $options = [];
-        $client = new GapicClientTraitStubExtension();
-        $updatedOptions = $client->call('buildClientOptions', [$options]);
+        $client = new StubGapicClientExtension();
+        $updatedOptions = $client->buildClientOptions($options);
 
         $this->assertArrayHasKey('addNewOption', $updatedOptions);
         $this->assertTrue($updatedOptions['disableRetries']);
     }
 
-    private function buildClientToTestModifyCallMethods()
+    private function buildClientToTestModifyCallMethods($clientClass = null)
     {
         $header = AgentHeader::buildAgentHeader([]);
         $retrySettings = $this->getMockBuilder(RetrySettings::class)
@@ -1151,7 +1146,8 @@ class GapicClientTraitTest extends TestCase
         $credentialsWrapper = CredentialsWrapper::build([
             'keyFile' => __DIR__ . '/testdata/json-key-file.json'
         ]);
-        $client = new GapicClientTraitStubExtension();
+        $clientClass = $clientClass ?: StubGapicClientExtension::class;
+        $client = new $clientClass();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -1188,12 +1184,12 @@ class GapicClientTraitTest extends TestCase
                 ])
             )
             ->willReturn(new FulfilledPromise(new Operation()));
-        $client->call('startCall', [
+        $client->startCall(
             'simpleMethod',
             'decodeType',
             [],
             new MockRequest(),
-        ])->wait();
+        )->wait();
     }
 
     public function testModifyUnaryCallFromOperationsCall()
@@ -1219,12 +1215,12 @@ class GapicClientTraitTest extends TestCase
                 ->disableOriginalConstructor()
                 ->setMethodsExcept(['validate'])
                 ->getMock();
-        $client->call('startOperationsCall', [
+        $client->startOperationsCall(
             'longRunningMethod',
             [],
             new MockRequest(),
             $operationsClient
-        ])->wait();
+        )->wait();
     }
 
     public function testModifyUnaryCallFromGetPagedListResponse()
@@ -1245,12 +1241,12 @@ class GapicClientTraitTest extends TestCase
                 ])
             )
             ->willReturn(new FulfilledPromise(new Operation()));
-        $client->call('getPagedListResponse', [
+        $client->getPagedListResponse(
             'pagedMethod',
             [],
             'decodeType',
             new MockRequest(),
-        ]);
+        );
     }
 
     /**
@@ -1274,7 +1270,7 @@ class GapicClientTraitTest extends TestCase
                 ])
             )
             ->willReturn($expectedResponse);
-        $client->call('startCall', $callArgs);
+        $client->startCall(...$callArgs);
     }
 
     public function modifyStreamingCallFromStartCallData()
@@ -1325,9 +1321,9 @@ class GapicClientTraitTest extends TestCase
     public function testGetTransport()
     {
         $transport = $this->getMockBuilder(TransportInterface::class)->getMock();
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('transport', $transport);
-        $this->assertEquals($transport, $client->call('getTransport'));
+        $this->assertEquals($transport, $client->getTransport());
     }
 
     public function testGetCredentialsWrapper()
@@ -1335,9 +1331,9 @@ class GapicClientTraitTest extends TestCase
         $credentialsWrapper = $this->getMockBuilder(CredentialsWrapper::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
         $client->set('credentialsWrapper', $credentialsWrapper);
-        $this->assertEquals($credentialsWrapper, $client->call('getCredentialsWrapper'));
+        $this->assertEquals($credentialsWrapper, $client->getCredentialsWrapper());
     }
 
     public function testUserProjectHeaderIsSetWhenProvidingQuotaProject()
@@ -1356,43 +1352,44 @@ class GapicClientTraitTest extends TestCase
                 $this->isInstanceOf(Call::class),
                 $this->equalTo([
                     'headers' => AgentHeader::buildAgentHeader([]) + [
-                        'X-Goog-User-Project' => [$quotaProject]
+                        'X-Goog-User-Project' => [$quotaProject],
+                        'User-Agent' => ['gcloud-php-legacy/']
                     ],
                     'credentialsWrapper' => $credentialsWrapper
                 ])
             );
-        $client = new GapicClientTraitStub();
-        $updatedOptions = $client->call('buildClientOptions', [
+        $client = new StubGapicClient();
+        $updatedOptions = $client->buildClientOptions(
             [
                 'transport' => $transport,
                 'credentials' => $credentialsWrapper,
             ]
-        ]);
-        $client->call('setClientOptions', [$updatedOptions]);
+        );
+        $client->setClientOptions($updatedOptions);
         $client->set('retrySettings', [
             'method' => $this->getMockBuilder(RetrySettings::class)
                 ->disableOriginalConstructor()
                 ->getMock()
             ]
         );
-        $client->call('startCall', [
+        $client->startCall(
             'method',
             'decodeType'
-        ]);
+        );
     }
 
     public function testDefaultScopes()
     {
-        $client = new GapicClientTraitDefaultScopeAndAudienceStub();
+        $client = new DefaultScopeAndAudienceGapicClient();
 
         // verify scopes are not set by default
-        $defaultOptions = $client->call('buildClientOptions', [[]]);
+        $defaultOptions = $client->buildClientOptions([]);
         $this->assertArrayNotHasKey('scopes', $defaultOptions['credentialsConfig']);
 
         // verify scopes are set when a custom api endpoint is used
-        $defaultOptions = $client->call('buildClientOptions', [[
+        $defaultOptions = $client->buildClientOptions([
             'apiEndpoint' => 'www.someotherendpoint.com',
-        ]]);
+        ]);
         $this->assertArrayHasKey('scopes', $defaultOptions['credentialsConfig']);
         $this->assertEquals(
             $client::$serviceScopes,
@@ -1400,10 +1397,10 @@ class GapicClientTraitTest extends TestCase
         );
 
         // verify user-defined scopes override default scopes
-        $defaultOptions = $client->call('buildClientOptions', [[
+        $defaultOptions = $client->buildClientOptions([
             'credentialsConfig' => ['scopes' => ['user-scope-1']],
             'apiEndpoint' => 'www.someotherendpoint.com',
-        ]]);
+        ]);
         $this->assertArrayHasKey('scopes', $defaultOptions['credentialsConfig']);
         $this->assertEquals(
             ['user-scope-1'],
@@ -1412,9 +1409,9 @@ class GapicClientTraitTest extends TestCase
 
         // verify empty default scopes has no effect
         $client::$serviceScopes = null;
-        $defaultOptions = $client->call('buildClientOptions', [[
+        $defaultOptions = $client->buildClientOptions([
             'apiEndpoint' => 'www.someotherendpoint.com',
-        ]]);
+        ]);
         $this->assertArrayNotHasKey('scopes', $defaultOptions['credentialsConfig']);
     }
 
@@ -1435,7 +1432,7 @@ class GapicClientTraitTest extends TestCase
             )
             ->shouldBeCalledOnce();
 
-        $client = new GapicClientTraitDefaultScopeAndAudienceStub();
+        $client = new DefaultScopeAndAudienceGapicClient();
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', []);
         $client->set(
@@ -1443,7 +1440,7 @@ class GapicClientTraitTest extends TestCase
             ['method.name' => $retrySettings->reveal()]
         );
         $client->set('transport', $transport->reveal());
-        $client->call('startCall', ['method.name', 'decodeType']);
+        $client->startCall('method.name', 'decodeType');
 
         $transport
             ->startUnaryCall(
@@ -1456,9 +1453,9 @@ class GapicClientTraitTest extends TestCase
             )
             ->shouldBeCalledOnce();
 
-        $client->call('startCall', ['method.name', 'decodeType', [
+        $client->startCall('method.name', 'decodeType', [
             'audience' => 'custom-audience',
-        ]]);
+        ]);
     }
 
     public function testDefaultAudienceWithOperations()
@@ -1490,7 +1487,7 @@ class GapicClientTraitTest extends TestCase
                 'totalPollTimeoutMillis' => 300,
             ]
         ];
-        $client = new GapicClientTraitDefaultScopeAndAudienceStub();
+        $client = new DefaultScopeAndAudienceGapicClient();
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', []);
         $client->set(
@@ -1505,12 +1502,12 @@ class GapicClientTraitTest extends TestCase
             ->getMock();
 
         // Test startOperationsCall with default audience
-        $client->call('startOperationsCall', [
+        $client->startOperationsCall(
             'method.name',
             [],
             new MockRequest(),
             $operationsClient,
-        ])->wait();
+        )->wait();
     }
 
     public function testDefaultAudienceWithPagedList()
@@ -1540,7 +1537,7 @@ class GapicClientTraitTest extends TestCase
                 'resourcesGetMethod' => 'getResources',
             ],
         ];
-        $client = new GapicClientTraitDefaultScopeAndAudienceStub();
+        $client = new DefaultScopeAndAudienceGapicClient();
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', []);
         $client->set(
@@ -1553,12 +1550,12 @@ class GapicClientTraitTest extends TestCase
         ]);
 
         // Test getPagedListResponse with default audience
-        $client->call('getPagedListResponse', [
+        $client->getPagedListResponse(
             'method.name',
             [],
             'decodeType',
             new MockRequest(),
-        ]);
+        );
     }
 
     public function testSupportedTransportOverrideWithInvalidTransport()
@@ -1566,29 +1563,29 @@ class GapicClientTraitTest extends TestCase
         $this->expectException(ValidationException::class);
         $this->expectExceptionMessage('Unexpected transport option "grpc". Supported transports: rest');
 
-        new GapicClientTraitRestOnly(['transport' => 'grpc']);
+        new RestOnlyGapicClient(['transport' => 'grpc']);
     }
 
     public function testSupportedTransportOverrideWithDefaultTransport()
     {
-        $client = new GapicClientTraitRestOnly();
+        $client = new RestOnlyGapicClient();
         $this->assertInstanceOf(RestTransport::class, $client->getTransport());
     }
 
     public function testSupportedTransportOverrideWithExplicitTransport()
     {
-        $client = new GapicClientTraitRestOnly(['transport' => 'rest']);
+        $client = new RestOnlyGapicClient(['transport' => 'rest']);
         $this->assertInstanceOf(RestTransport::class, $client->getTransport());
     }
 
     /** @dataProvider provideDetermineMtlsEndpoint */
     public function testDetermineMtlsEndpoint($apiEndpoint, $expected)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         $this->assertEquals(
             $expected,
-            $client->call('determineMtlsEndpoint', [$apiEndpoint])
+            $client->determineMtlsEndpoint($apiEndpoint)
         );
     }
 
@@ -1614,12 +1611,12 @@ class GapicClientTraitTest extends TestCase
      */
     public function testShouldUseMtlsEndpoint($envVarValue, $options, $expected)
     {
-        $client = new GapicClientTraitStub();
+        $client = new StubGapicClient();
 
         putenv('GOOGLE_API_USE_MTLS_ENDPOINT=' . $envVarValue);
         $this->assertEquals(
             $expected,
-            $client->call('shouldUseMtlsEndpoint', [$options])
+            $client->shouldUseMtlsEndpoint($options)
         );
     }
 
@@ -1646,8 +1643,8 @@ class GapicClientTraitTest extends TestCase
             putenv($envVar);
         }
 
-        $client = new GapicClientTraitStub();
-        $options = $client->call('buildClientOptions', [$options]);
+        $client = new StubGapicClient();
+        $options = $client->buildClientOptions($options);
 
         // Only check the keys we care about
         $options = array_intersect_key(
@@ -1711,18 +1708,155 @@ class GapicClientTraitTest extends TestCase
         putenv('GOOGLE_API_USE_MTLS_ENDPOINT=auto');
         putenv(CredentialsLoader::MTLS_CERT_ENV_VAR . '=true');
 
-        $client = new GapicClientTraitStub();
-        $options = $client->call('buildClientOptions', [[]]);
+        $client = new StubGapicClient();
+        $options = $client->buildClientOptions([]);
 
         $this->assertSame('test.mtls.address.com:443', $options['apiEndpoint']);
         $this->assertTrue(is_callable($options['clientCertSource']));
         $this->assertEquals(['foo', 'foo'], $options['clientCertSource']());
     }
+
+    public function testInvalidClientOptionsTypeThrowsExceptionForV2SurfaceOnly()
+    {
+        // v1 client
+        new StubGapicClient(['apiEndpoint' => ['foo']]);
+        $this->assertTrue(true, 'Test made it to here without throwing an exception');
+
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage(
+            PHP_MAJOR_VERSION < 8
+                ? 'Argument 1 passed to Google\ApiCore\Options\ClientOptions::setApiEndpoint() '
+                    . 'must be of the type string or null, array given'
+                : 'Google\ApiCore\Options\ClientOptions::setApiEndpoint(): Argument #1 '
+                    . '($apiEndpoint) must be of type ?string, array given'
+        );
+
+        // v2 client
+        new GapicV2SurfaceClient(['apiEndpoint' => ['foo']]);
+    }
+
+    public function testCallOptionsForV2Surface()
+    {
+        list($client, $transport) = $this->buildClientToTestModifyCallMethods(
+            GapicV2SurfaceClient::class
+        );
+
+        $transport->expects($this->once())
+            ->method('startUnaryCall')
+            ->with(
+                $this->isInstanceOf(Call::class),
+                $this->equalTo([
+                    'headers' => AgentHeader::buildAgentHeader([]) + ['Foo' => 'Bar'],
+                    'credentialsWrapper' => CredentialsWrapper::build([
+                        'keyFile' => __DIR__ . '/testdata/json-key-file.json'
+                    ]),
+                    'timeoutMillis' => null, // adds null timeoutMillis
+                ])
+            )
+            ->willReturn(new FulfilledPromise(new Operation()));
+
+        $callOptions = [
+            'headers' => ['Foo' => 'Bar'],
+            'invalidOption' => 'wont-be-passed'
+        ];
+        $client->startCall(
+            'simpleMethod',
+            'decodeType',
+            $callOptions,
+            new MockRequest(),
+        )->wait();
+    }
+
+    public function testInvalidCallOptionsTypeForV1SurfaceDoesNotThrowException()
+    {
+        list($client, $transport) = $this->buildClientToTestModifyCallMethods();
+
+        $transport->expects($this->once())
+            ->method('startUnaryCall')
+            ->with(
+                $this->isInstanceOf(Call::class),
+                $this->equalTo([
+                    'transportOptions' => ['custom' => ['addModifyUnaryCallableOption' => true]],
+                    'headers' => AgentHeader::buildAgentHeader([]),
+                    'credentialsWrapper' => CredentialsWrapper::build([
+                        'keyFile' => __DIR__ . '/testdata/json-key-file.json'
+                    ]),
+                    'timeoutMillis' => 'blue', // invalid type, this is ignored
+                ])
+            )
+            ->willReturn(new FulfilledPromise(new Operation()));
+
+        $client->startCall(
+            'simpleMethod',
+            'decodeType',
+            ['timeoutMillis' => 'blue'],
+            new MockRequest(),
+        )->wait();
+    }
+
+    public function testInvalidCallOptionsTypeForV2SurfaceThrowsException()
+    {
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage(
+            PHP_MAJOR_VERSION < 8
+                ? 'Argument 1 passed to Google\ApiCore\Options\CallOptions::setTimeoutMillis() '
+                    . 'must be of the type int or null, string given'
+                : 'Google\ApiCore\Options\CallOptions::setTimeoutMillis(): Argument #1 '
+                    . '($timeoutMillis) must be of type ?int, string given'
+        );
+
+        list($client, $_) = $this->buildClientToTestModifyCallMethods(GapicV2SurfaceClient::class);
+
+        $client->startCall(
+            'simpleMethod',
+            'decodeType',
+            ['timeoutMillis' => 'blue'], // invalid type, will throw exception
+            new MockRequest(),
+        )->wait();
+    }
+
+    public function testSurfaceAgentHeaders()
+    {
+        // V1 does not contain new headers
+        $client = new RestOnlyGapicClient([
+            'gapicVersion' => '0.0.2',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.2 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-legacy/0.0.2', $agentHeader['User-Agent'][0]);
+
+        // V2 contains new headers
+        $client = new GapicV2SurfaceClient([
+            'gapicVersion' => '0.0.1',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.1 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-new/0.0.1', $agentHeader['User-Agent'][0]);
+    }
 }
 
-class GapicClientTraitStub
+class StubGapicClient
 {
-    use GapicClientTrait;
+    use GapicClientTrait {
+        buildClientOptions as public;
+        buildRequestParamsHeader as public;
+        configureCallConstructionOptions as public;
+        createCredentialsWrapper as public;
+        createOperationsClient as public;
+        createTransport as public;
+        determineMtlsEndpoint as public;
+        getGapicVersion as public;
+        getCredentialsWrapper as public;
+        getPagedListResponse as public;
+        getTransport as public;
+        setClientOptions as public;
+        shouldUseMtlsEndpoint as public;
+        startApiCall as public;
+        startAsyncCall as public;
+        startCall as public;
+        startOperationsCall as public;
+    }
+    use GapicClientStubTrait;
 
     public static function getClientDefaults()
     {
@@ -1743,12 +1877,10 @@ class GapicClientTraitStub
             ],
         ];
     }
+}
 
-    public function call($fn, array $args = [])
-    {
-        return call_user_func_array([$this, $fn], $args);
-    }
-
+trait GapicClientStubTrait
+{
     public function set($name, $val, $static = false)
     {
         if (!property_exists($this, $name)) {
@@ -1770,7 +1902,7 @@ class GapicClientTraitStub
     }
 }
 
-class GapicClientTraitStubExtension extends GapicClientTraitStub
+class StubGapicClientExtension extends StubGapicClient
 {
     protected function modifyClientOptions(array &$options)
     {
@@ -1801,9 +1933,15 @@ class GapicClientTraitStubExtension extends GapicClientTraitStub
     }
 }
 
-class GapicClientTraitDefaultScopeAndAudienceStub
+class DefaultScopeAndAudienceGapicClient
 {
-    use GapicClientTrait;
+    use GapicClientTrait {
+        buildClientOptions as public;
+        startCall as public;
+        startOperationsCall as public;
+        getPagedListResponse as public;
+    }
+    use GapicClientStubTrait;
 
     const SERVICE_ADDRESS = 'service-address';
 
@@ -1821,28 +1959,14 @@ class GapicClientTraitDefaultScopeAndAudienceStub
             ],
         ];
     }
-
-    public function call($fn, array $args = [])
-    {
-        return call_user_func_array([$this, $fn], $args);
-    }
-
-    public function set($name, $val, $static = false)
-    {
-        if (!property_exists($this, $name)) {
-            throw new \InvalidArgumentException("Property not found: $name");
-        }
-        if ($static) {
-            $this::$$name = $val;
-        } else {
-            $this->$name = $val;
-        }
-    }
 }
 
-class GapicClientTraitRestOnly
+class RestOnlyGapicClient
 {
-    use GapicClientTrait;
+    use GapicClientTrait {
+        buildClientOptions as public;
+        getTransport as public;
+    }
 
     public function __construct($options = [])
     {
@@ -1865,16 +1989,6 @@ class GapicClientTraitRestOnly
         ];
     }
 
-    public function call($fn, array $args = [])
-    {
-        return call_user_func_array([$this, $fn], $args);
-    }
-
-    public function getTransport()
-    {
-        return $this->transport;
-    }
-
     private static function supportedTransports()
     {
         return ['rest', 'fake-transport'];
@@ -1884,9 +1998,14 @@ class GapicClientTraitRestOnly
     {
         return 'rest';
     }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
+    }
 }
 
-class GapicClientTraitOperationsStub extends GapicClientTraitStub
+class OperationsGapicClient extends StubGapicClient
 {
     public $operationsClient;
 
@@ -1900,5 +2019,39 @@ class CustomOperationsClient
 {
     public function getOperation($name, $arg1, $arg2)
     {
+    }
+}
+
+class GapicV2SurfaceClient
+{
+    use GapicClientTrait {
+        startCall as public;
+    }
+    use GapicClientStubTrait;
+
+    public function __construct(array $options = [])
+    {
+        $clientOptions = $this->buildClientOptions($options);
+        $this->setClientOptions($clientOptions);
+    }
+
+    public static function getClientDefaults()
+    {
+        return [
+            'apiEndpoint' => 'test.address.com:443',
+            'serviceName' => 'test.interface.v1.api',
+            'clientConfig' => __DIR__ . '/testdata/test_service_client_config.json',
+            'descriptorsConfigPath' => __DIR__.'/testdata/test_service_descriptor_config.php',
+            'transportConfig' => [
+                'rest' => [
+                    'restClientConfigPath' => __DIR__.'/testdata/test_service_rest_client_config.php',
+                ]
+            ],
+        ];
+    }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
     }
 }